fix(universe): compose OnState(StateActive) instead of clobbering
Pre-existing regression introduced in 7bbbf3b — the unconditional
universe-side OnState(StateActive) registration in createNode was
last-writer-wins over any callbacks the world factory registered.
NewGameWorld in the space game registers its OnEnter via
gw.Players.OnState(StateActive) with spawn logic; the universe block
ran AFTER the world factory and silently overwrote it. Result: every
space-game login completed but never spawned a player entity.

The 4node-basic example sidesteps this by registering its spawn via
coord.OnPlayerJoin (which feeds the universe-side joinHooks), so the
break only manifested in games that wire OnState directly.

Fix: snapshot any prior callbacks via PlayerManager.StateCallbacks
(new accessor) before registering, then chain prior.OnEnter and
prior.OnExit inside the new fan-out — debug-flag hydration runs
first, then prior OnEnter (game spawn), then process-level joinHooks.
OnExit runs default cleanup, then prior OnExit, then leaveHooks.

+++ cmd/server/main.go
@@ -162,7 +162,7 @@ func main() {
 
 	coordCfg.Logger = gameLog
 	coordCfg.LoginHandler = mmokit.HandleLogin(
-		uint32(enginepb.ClientEventCode_CE_LOGIN),
+		enginepb.ClientEventCode_CE_LOGIN,
 		func(m *enginepb.LoginMsg) (string, any, error) {
 			name, err := mmokit.ValidateUsername(m.Username, 0)
 			return name, nil, err
